requirements.txt
python-telegram-bot[job-queue]==21.0
google-generativeai>=0.7.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pillow==11.0.0
apscheduler==3.10.4
